refactor: move middleware to src/middleware.ts
Merge index.ts and utils.ts into a single file.
