+++ src/middleware.ts
@@ -1,4 +1,6 @@
+import { authenticate } from "@navikt/astro-auth";
 import type { APIContext } from "astro";
+import { defineMiddleware, sequence } from "astro/middleware";
 
 const BASE = "/dokumentarkiv";
 const LOCALES = ["nb", "nn", "en"];
@@ -7,7 +9,7 @@ const DEFAULT_LOCALE = "nb";
 const isInternal = (context: APIContext) =>
   context.url.pathname.includes("/internal/");
 
-export const defaultLocaleRedirect = (context: APIContext): string | null => {
+const defaultLocaleRedirect = (context: APIContext): string | null => {
   const { pathname, search } = context.url;
 
   if (!pathname.startsWith(BASE)) {
@@ -28,3 +30,14 @@ export const defaultLocaleRedirect = (context: APIContext): string | null => {
   const localizedPath = `${BASE}/${DEFAULT_LOCALE}${rest === "" ? "/" : rest}`;
   return `${localizedPath}${search}`;
 };
+
+const localeMiddleware = defineMiddleware(async (context, next) => {
+  const localeRedirect = defaultLocaleRedirect(context);
+  if (localeRedirect) {
+    return context.redirect(localeRedirect);
+  }
+
+  return next();
+});
+
+export const onRequest = sequence(authenticate(), localeMiddleware);
